requirements.txt
user-agent
requests
beautifulsoup4
python-dotenv
datetime
colorama

